+++ packages/discovery/src/discovery/proxies/ProxyDetector.test.ts
@@ -72,6 +72,7 @@ describe(ProxyDetector.name, () => {
       'Eternal Storage proxy': async () => undefined,
       'Polygon Extension proxy': async () => undefined,
       'Optics Beacon proxy': async () => undefined,
+      immutable: async () => undefined,
     } as Record<ManualProxyType, Detector>)
     const result = await detector.detectProxy(
       EthereumAddress.random(),