+++ packages/discovery/src/discovery/proxies/ProxyDetector.ts
@@ -14,6 +14,7 @@ import { detectStarkWareProxy } from './auto/StarkWareProxy'
 import { detectZeppelinOSProxy } from './auto/ZeppelinOSProxy'
 import { getCallImplementationProxy } from './manual/CallImplementationProxy'
 import { getEternalStorageProxy } from './manual/EthernalStorageProxy'
+import { getImmutableProxy } from './manual/immutableProxy'
 import { getNewArbitrumProxy } from './manual/NewArbitrumProxy'
 import { getOpticsBeaconProxy } from './manual/OpticsBeaconProxy'
 import { getPolygonExtensionProxy } from './manual/PolygonExtensionProxy'
@@ -47,6 +48,7 @@ const MANUAL_DETECTORS: Record<ManualProxyType, Detector> = {
   'Eternal Storage proxy': getEternalStorageProxy,
   'Polygon Extension proxy': getPolygonExtensionProxy,
   'Optics Beacon proxy': getOpticsBeaconProxy,
+  immutable: getImmutableProxy,
 }
 
 export class ProxyDetector {